Remove uses of REPONAME placeholder text.
