+++ templates/benchmarks/asv.conf.json
@@ -3,7 +3,7 @@
     "project": "REPONAME according to pip, e.g. scitools-iris",
     "project_url": "https://github.com/SciTools/REPONAME",
     "repo": "..",
-    "environment_type": "delegated-REPONAME",
+    "environment_type": "delegated",
     "show_commit_url": "https://github.com/scitools/REPONAME/commit/",
     "branches": ["upstream/main"],
 
@@ -11,7 +11,7 @@
     "env_dir": ".asv/env",
     "results_dir": ".asv/results",
     "html_dir": ".asv/html",
-    "plugins": [".asv_delegated_REPONAME"],
+    "plugins": [".asv_delegated"],
 
     "command_comment": [
         "Can get some speedup by skipping project build, install and ",
