+++ templates/benchmarks/custom_bms/tracemallocbench.py
@@ -1,6 +1,6 @@
-# Copyright REPONAME contributors
+# Copyright SciTools contributors
 #
-# This file is part of REPONAME and is released under the BSD license.
+# This file is part of SciTools and is released under the BSD license.
 # See LICENSE in the root of the repository for full licensing details.
 
 """Benchmark for growth in process resident memory, repeating for accuracy.
